Latex removed and markdown and code files made simpler
