+++ front/package.json
@@ -24,6 +24,7 @@
     "katex": "^0.16.21",
     "lucide-react": "^0.474.0",
     "mathjax-full": "^3.2.2",
+    "mermaid": "^11.4.1",
     "papaparse": "^5.5.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
